Import
This program was written to calculate the distances between NBA cities.
Queries were sent to Wolfram-Alpha which returned the distance in miles
between cities. This was done with each combination of NBA teams to
find out which teams are farthest away from each other. See the text
file for the results.